Source Code/Source Code/requirements.txt
customtkinter>=5.2.0
pymodbus>=3.6.0
pyserial>=3.5
Pillow>=10.0.0
pyinstaller>=6.0.0
